cicada2/runners/kafka_runner/requirements.txt
kafka-python==2.0.1
grpcio
protobuf
typing_extensions
